vector improvements, mat4f

+++ src/main/java/de/bixilon/minosoft/gui/rendering/gui/elements/input/button/AbstractButtonElement.kt
@@ -181,7 +181,7 @@ abstract class AbstractButtonElement(
     override fun onChildChange(child: Element) {
         if (child == textElement) {
             if (dynamicSized) {
-                size = textElement.size + Vec2i(TEXT_PADDING * 2, TEXT_PADDING * 2)
+                size = textElement.size + Vec2f(TEXT_PADDING * 2, TEXT_PADDING * 2)
             }
             cacheUpToDate = false
         }
